refactor: extract import writing and env-merge to lib/importIO.ts
Move writeImportedFiles, writeImportedEnvironmentFile, the discovered-
variable merge (applyImportedVariables) and the per-format import
dispatch (importCollectionContent) out of App.svelte into
src/lib/importIO.ts. App.svelte keeps only the modal visibility state
and shows the env-target modal for variables the importer returns.
No behavior change.
